python-mysql/requirements.txt
PyMySQL
mysql-connector-python